+++ src/views-components/move-to-dialog/move-to-dialog.tsx
@@ -6,7 +6,7 @@ import * as React from "react";
 import { InjectedFormProps, Field } from 'redux-form';
 import { WithDialogProps } from '~/store/dialog/with-dialog';
 import { FormDialog } from '~/components/form-dialog/form-dialog';
-import { ProjectTreePicker } from '~/components/project-tree-picker/project-tree-picker';
+import { ProjectTreePickerField } from '~/views-components/project-tree-picker/project-tree-picker';
 import { MOVE_TO_VALIDATION } from '~/validators/validators';
 import { MoveToFormDialogData } from '~/store/move-to-dialog/move-to-dialog';
 
@@ -21,6 +21,6 @@ export const MoveToFormDialog = (props: WithDialogProps<string> & InjectedFormPr
 const MoveToDialogFields = () =>
     <Field
         name="ownerUuid"
-        component={ProjectTreePicker}
+        component={ProjectTreePickerField}
         validate={MOVE_TO_VALIDATION} />;
