+++ src/views-components/project-copy-project-dialog/project-copy-project-dialog.tsx
@@ -4,7 +4,7 @@
 
 import { compose } from "redux";
 import { withDialog } from "~/store/dialog/with-dialog";
-import { reduxForm, startSubmit, stopSubmit } from 'redux-form';
+import { reduxForm } from 'redux-form';
 import { PROJECT_COPY_DIALOG } from '~/store/project-copy-project-dialog/project-copy-project-dialog';
 import { ProjectCopyFormDialog } from "~/views-components/project-copy-dialog/project-copy-dialog";
 import { copyProject } from '../../store/project-copy-project-dialog/project-copy-project-dialog';
